Move SQS under aws/ with canonical shared AWS config
Reorganize SQS from queue/sqs/ to aws/sqs/ across all three layers
(api, service, boot), matching the existing S3 organization pattern.

SQS now acquires credentials through the shared config.aws resource
(service/aws/config/manager.go) via the resource registry, removing
inline AWS credential fields, TLS, HTTP transport, and custom JSON
marshaling from the config.
