+++ boot/components/queue/constants.go
@@ -10,6 +10,5 @@ const (
 	ConsumersName    boot.Name = "queue-consumers"
 	MemoryDriverName boot.Name = "queue-memory"
 	AMQPDriverName   boot.Name = "queue-amqp"
-	SQSDriverName    boot.Name = "queue-sqs"
 	RedisDriverName  boot.Name = "queue-redis"
 )
