+++ boot/components/service/aws/constants.go
@@ -7,4 +7,5 @@ import "github.com/wippyai/runtime/api/boot"
 const (
 	ConfigName boot.Name = "aws_config"
 	S3Name     boot.Name = "s3"
+	SQSName    boot.Name = "sqs_queue"
 )
