+++ service/aws/sqs/util.go
